Issue#37: Moved REST code into separate package
